13.40.1 - remove hard error on VJOY device hidden
Removes forced hard error if a VJOY device is not available to the code.
This can occur if the GremlinEX process is not whitelisted by HIDHide or if another process is running and has a hard lock on a VJOY device.

This may not fix conflicts but prevents a hard error.