10.40.3 Potential fix for adding new mode crashing QT on mode change
Tentative fix for mode change crash with no errors.

Prevented a QT UI update event from firing when the mode changes which appears to crash QT6 internally.   If QT crashes internally - it takes Python down with it which is why there is no error.

This is a tentative fix subject to more testing to get things going now in case the fix isn't complete but to get folks going.