m76T115
### (m76T115)
- New: (experimental) Enhanced filters for joystick input devices (see below).
 Fix: Added more manual QT widget life cycle management logic rather than rely on QT's internal management mechanisms prone to memory leaks and synchronization issues between the Python memory manager and the QT C++ memory manager.
- Fix: OSC sorting key error when sorting messages with, or without data
- Fix: axis state module now handles OSC axis inputs properly for OSC linear data

**Input Filtering**

GEX will now filter (limit) the number of inputs shown on each joystick device to a reasonable and more manageable number.  This is directly aimed at performance by reducing the number of entities visible at anyone time.  It is also for convenience and ease of use reasons: until now, a large number of devices and inputs could significantly impact UI performance and memory at profile design time, whether relevant or not.  These could consume thousands of entities that need to be tracked with a noticeable impact on performance and responsiveness.

Starting with this patch, each joystick device can easily hide/show inputs. Only visible inputs are tracked which significantly reduces memory and CPU usage at design time.

Inputs are filtered on a per joystick device basis.  Each joystick device's inputs can be shown or hidden via the new device filter options dialog.  This dialog is accessible from the filter button above the device input list.

The filter settings are per profile and persisted in the profile so the profile must be saved for these settings to persist from one session to another.

A new filter option page is also available in the global options dialog to set these initial default limits. The options page also has global buttons to apply the filter on all joystick devices in the profile, although these are duplicated in the per device filter dialog via control-click on the shortcut.

If a legacy profile is loaded with no filtering information, GEX will apply filtering to show mapped inputs only.

New profiles will used global filtering defaults as set in global options (4 axes, 8 buttons and 1 hat).

"Reasonable" varies with the number of devices, inputs and hardware capabilities. Performance is directly impacted by the number of visible inputs, so it is highly recommended to keep those to a minimum. Inputs can easily be hidden or shown via the filter dialog.  Hiding an input does not delete its mappings.  Mappings however can only be seen if the input is visible.

With this new performance oriented mechanism, the input your are looking for may not be visible depending on how many devices you have mapped and default filtering applied.  Existing profiles with hundreds of mappings may still be visible by default until changed, so may still impact performance negatively until the number of displayed inputs is reduced.

The filtering does not impact profile execution.  Invisible inputs will still run their mappings if the input is mapped.

Filters only apply to joystick devices (physical and virtual) as they usually represent the bulk of input counts.

Please remember to save the profile to have the filters persist as the data is not saved until the profile is saved.

A note on performance:  The filtering mechanism reduces the memory footprint of GEX significantly, and increases UI responsiveness significantly as well on most systems.